refactor: Migrate to v3 managers as default implementation
BREAKING CHANGE: V3 managers are now the default implementation

- Moved old managers to legacy/ directory for reference
  - legacy/db_manager_old.py (old DatabaseManager)
  - legacy/channel_manager_old.py (old ChannelManager)
  - legacy/schema_old.sql (old database schema)

- Renamed v3 managers to be the default (removed V3 suffix):
  - DatabaseManagerV3 → DatabaseManager
  - ChannelManagerV3 → ChannelManager
  - AgentManagerV3 → AgentManager
  - SubscriptionManagerV3 → SubscriptionManager
  - schema_v3.sql → schema.sql

- Updated all imports and class references throughout codebase
- Updated test files to use new names
- All tests passing ✅

This completes the Phase 2 migration to the unified permission system.
The v3 architecture is now the standard implementation with:
- Database-driven permissions
- Proper separation of concerns
- Unified channel system
- Agent DM policies and discovery
